move parse_autocomplete_message to metadata mod
